Refactored the godly Compiler class into multiple Build Tool and Utilities classes.
