+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "instapack",
-  "version": "5.0.0-rc1",
+  "version": "5.0.0-rc2",
   "description": "All-in-one web app client project compiler using TypeScript and Sass!",
   "main": "bin/index.js",
   "bin": {
@@ -37,7 +37,7 @@
     "thread-loader": "1.1.2",
     "ts-loader": "3.1.0",
     "typescript": "2.6.0-rc",
-    "uglify-es": "3.1.5",
+    "uglify-es": "3.1.6",
     "uglifyjs-webpack-plugin": "1.0.1",
     "undertaker": "1.2.0",
     "vue-template-compiler": "2.5.2",
